+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add FreeBSD on arm64 target #29
 - Add the `darwin-image` command to build the darwin docker image
+- Add the `local` flag for darwin to build directly from the host
 - Add a dedicated docker image for macOS
 - Add a dedicated docker image for Windows